requirement.txt
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@2bd65027ca5c3b87b77d4f05b8eacae58d8d106f